[Update] WKLAY address moved to config file
