+++ subgraphs/exchange/src/utils/config.ts
@@ -1,0 +1,2 @@
+export const WKLAY_ADDRESS = "0x73365f8f27de98d7634be67a167f229b32e7bf6c";
+export const KlayOracleAddress = "0xeD074DA2A76FD2Ca90C1508930b4FB4420e413B0";
